backend/pyproject.toml
[project]
name = "book-content-rag-pipeline"
version = "0.1.0"
description = "RAG Pipeline for Book Content: Crawling, Chunking, Embedding, and Vector Storage"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=5.5.0",
    "qdrant-client>=1.8.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "pydantic>=2.5.0",
]
